Add stat tracking to credits.
We can now print any 1, 2, or 3 byte integer out of accessible
memory with an escape sequence in DrawComplexString. However it
clobbers the party's gold so it can only be used in the credits.

The sequence is 1010XXADDR04 where XX is the number of bytes and
ADDR is a two byte pointer to the start of the integer. The last
04 is the typical code to print the party's current GP. Gold was
chosen because it's already 3 bytes and can handle anything else.

Updated credits and story pages with some stats to use this new
feature, including gold, steps taken, soft resets and hard boots.