Flags rework. Bump version to 1.6.1.
Fixed a bug where TeamSteak flag and PartyLeaderShuffle were getting
mixed up, but also overhauled flags so they are not limited to 32 bits
anymore. I reordered a couple of flags to match the JavaScript limitation
of only having the non-binary flags at the end.

Continued using our variant on base64, but use the real thing inside
the javascript code as long as possible to leverage atob and btoa.
However I did remove the trimming of the = characters to keep it more
legit, and add a - between the base64 bits and the slider data.