Dynamic Menu BG Color
With select button handling contributions from MeridianBC. The select
button on the opening menu will now increment the bgcolor byte. It is
kept in unsram and saved along with everything else. Power cycles show
blue until the save file is loaded (like classic FF that had save).

Overhaul LoadBorderPalette_Blue into a method that dynamically loads the
color before calling out of bank to a large method that writes the palette
including the dynamic color. This method edits both the in room and out of
room palettes.

Fixed a couple equipment screen bugs that use palette 1 for no reason. In
vanilla it's the same as palette 03 but not anymore.