Non-incentivized shop fix.
The shop item was being removed in the wrong spot, and potentially
twice, but not when it came from the treasurePool. That's actually
when the failure occured because there would be too many items and
not enough locations.

Tidied up a bunch of PlaceSaneItems to utilize LINQ a little more
and generally leave things in a better way then I found them.