Fix ToFR always appearing on the second floor.
Floor shuffle was a little bugged about a few dead ends that would
always appear as second floors and ToFR was one of them. I'll add
another follow up to enable ToFR at the bottom of a dungeon but for
now it will be always placed at ground floor like Ordeals and
Coneria Castle.