Add basic support for overrides
The idea of an "override" is to get the randomizer to create certain
items, or place certain things in specific places. (Or to put them in
places they normally wouldn't end up.)

This first change only deals with magic shuffle, to put spells in
particular slots, and item magic, so particular magic items are turned
into specific other items, such as a "Lamp Shirt" or "Nuke Rod".