Support text seeds.
This adds support for seeds to not be limited to 8 hexadecimal digits,
expanding it to support up to 8 "characters".

This is actually up to 8 bytes worth of characters, so while unicode
characters work, it depends on how many bytes they take.

To help distinguish between hex and text seeds, the seed box shows green
for hex seeds and blue for text.