Fix Circle CI.
Because "CIRCLE_BRANCH" is the name of the branch being built, and
because github PRs are done from branches such as "pull/1", the if/else
logic to check for building alpha doesn't work.

Since this file is in the alpha branch, and alpha no longer includes the
FF1RandomizerOnline project, it makes sense to remove it from the build
step at least.