Increase thief starting agility, agility growth, and starting evade.
This is to make it more viable as a first-slot character and
compensate for its poor HP, absorb and mdef compared to the fighter
and black belt.

A starting AGI of 120 was selected after a lot of spread sheet number
crunching and a few test seeds at various agility values.

This gives the thief a starting evade (E) of 168.  Because the base
to-hit value is also 168, and the to-hit calculation is (168-E+H),
this means the monster's accuracy value (H) dominates in determining
if how easily the thief can be hit.  Vanilla monster accuracy ranges
from 2 (imps) to 114 (Kraken 2) to 200
(Chaos), with the majority of monsters having
accuracy in the 20-80 range.

This means a monster with an accuracy of 50 should have about a 25%
chance of hitting the thief.

HOWEVER monster accuracy gets scaled along with all the other monster
stats, so in practice most of the more dangerous monsters don't have
much trouble landing blows.

Another implication for combat is that enemies that have multiple
attacks (most notabliy Kary and Kraken) will tend to miss _some_ hits,
but not all.  This compensates for the fact that the ninja has garbage
HP growth.  A knight with 600 HP can tank 400 HP of damage and
survive, a ninja with 400 HP can evade half the blows and take 200 HP
of damage, with the same outcome.

The agility stat, together with the luck stat, plays a role in
determining whether the party is "suprised" or gets "first strike".
This is called "initiative".  The calculation also incorporates a
"suprise factor" of the monster formation.  For most formations, the
suprise factor is 4.  A first-slot thief will have about a 25% chance
of getting first strike.  This sounds like a lot, except a) some
formations have higher suprise factors.  b) you can't first strike
unrunnable formations
(unless you enable the
UnrunnablesStrikeFirstAndSurprise flag)

For formations with a higher surprise factor, the thief has a
smaller (or even zero) chance to get first strike -- but the thief's
initiative also makes also less likely (or zero) that the party will
be surprised.  This turns out to be a subtle but important
contribution to surviability, by reducing the incidence of
party-wiping ambushes.

Overall, having playing a test seed with a lead thief, it didn't feel
overpowered.

a) This only reduces damage taken from basic melee.  The thief remains
as vulnerable as ever to spells, sleep/stun, and still has lousy HP
growth and (prior to class change) weapon and armor selection compared
to the knight.

b) The buff to the suprise/first stike calculation is mainly useful
for avoiding combat, not overcoming difficult battles.

That said, it's possible this could be somewhat more unbalancing with
3 warrior or 2 warrior teams -- we can always dial this back or add
options based on community feedback.