Place shards separately from leftoverItems
Adjust item placement so LooseExcludePlacedDungeons does not apply to
shards.  This fixes a couple of problems:

1) Shards will be placed anywhere, instead of only non-incentive
locations (which sharply limited shard placement in a way that was not
intended.)

2) Fixes a crash when LooseExcludePlacedDungeons is enabled along with
every incentive item, every incentive location, and shard
hunt -- because previously, there would be more shards+loose items to
place than available non-incentive locations).