Tetron/selectable thief agi (#610)
Make thief starting AGL selectable between 80, 100, 120

This gives some additional balance options in case the community feels that AGL 120 (which hits the stat cap at level 32) is excessive.

Starting AGL 100 gives a starting evade of 148.  A level 50 thief/ninja will get to evade of 198 (almost but not quite the stat cap of 200).

Starting AGL 80 gives a starting evade of 128.  A level 50 thief will get to evade of 178.

Also: always apply the 3 digit stats bugfix.  Remove it as a separate flag to reclaim one bit of the flags string.