Replace floater in unincentivizedQuestItems list when IsFloaterRemoved
This fixes a major placement bug in the previous logic:

The floater would get placed as a loose item, progression items would
get placed on the basis of assumed access to the airship, but after
sane placement was complete, the the floater was replaced with a dud
item (cabin), making progression impossible.

This removes the floater entirely, prior to placement, as intended.