Improve versioned deployment to get more stable URLs (#636)
* Improve versioned deployment to get more stable URLs

The goal of this proposed change is to ensure stable URLs by always
sending the user to an instanced site.  This way, when the user copies
a link out of the URL bar to share with others, they get a stable URL.

1) instanced master deploy is the same x-y-z.ffr.com

2) beta site now deploys as beta-<commit>.ffr.com

3) ffr.com and beta.ffr.com front page deploy a simple index.html with
a meta refresh redirect to the most recently deployed instanced site.