Made RomStampModel handle hex and added constants
RomStampModel now contains contants for its bank and starting offset. I also added a method to create a hex string of its contents and convert from Hex back to ASCII. I also made a mistake in how long the flag string was, so I updated that and the filler amount. The RomStampModel will now occupy 350 bytes of data total