Fixing the following soft lock situations:
1. The region where the airship is placed won't be processed by
PreventAirshipLanding.  To avoid a situation where the region where
the airship appears is already 100% covered by non-landable tiles, the
airship feature now has two land squares.  This prevents a soft lock
whereby a player docks the ship, gets the airship, flys away, has
required progression in the same region as the airship, but can't
land.

2. Added a missing edge border to the the E/W canal features.
Previously they could be placed in locations with water on the other
side, giving the ship unintended access to the dock.  Not only does
this subvert the intended progression, it is possible for the canal to
be placed in the canal-locked region, leading players to think they
are soft locked.